make RxJava helpers optional in Inserter
